Initial commit - ran create-react-app with typescript
I've used the create-react-app since it creates the skeleton app, preconfigured with linting, transpiling, hot reloading and unit testing built in. To get going quickly. I decided to use the typescript template so I can make the application more readable and robust.

This will serve as a great base to start the project.